0328 목요일 이혜원 reference.js / approval.js 삭제 draftDetail 추가, 나머지 수정
