+++ GreenLight/src/main/java/com/green/light/model/mapper/ApprovalDaoImpl.java
@@ -7,6 +7,8 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Repository;
 
 import com.green.light.vo.ApprJstreeVo;
+import com.green.light.vo.DepartmentVo;
+import com.green.light.vo.DocumentVo;
 import com.green.light.vo.EmployeeVo;
 
 @Repository
@@ -15,15 +17,9 @@ public class ApprovalDaoImpl implements IApprovalDao{
 	private SqlSessionTemplate sqlSession;
 	private final String NS = "com.green.light.model.mapper.ApprovalDaoImpl.";
 	
-	@Override
-	public EmployeeVo selectDrafterDeptMgr(String id) {
-		return sqlSession.selectOne(NS+"selectDrafterDeptMgr", id);
-	}
 	
-	@Override
-	public EmployeeVo selectDeptMgrByDept(String deptno) {
-		return sqlSession.selectOne(NS+"selectDeptMgrByDept", deptno);
-	}
+	
+	
 	
 	@Override
 	public EmployeeVo selectCEO() {
