+++ GreenLight/src/main/java/com/green/light/model/mapper/DepartmentDaoImpl.java
@@ -8,6 +8,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Repository;
 
 import com.green.light.vo.DepartmentVo;
+import com.green.light.vo.EmployeeVo;
 
 @Repository
 public class DepartmentDaoImpl implements IDepartmentDao {
@@ -65,6 +66,16 @@ public class DepartmentDaoImpl implements IDepartmentDao {
 	@Override
 	public int updateDeptMgr(Map<String, Object> map) {
 		return sqlSession.update(NS+"updateDeptMgr", map);
+	}
+
+	@Override
+	public DepartmentVo selectDeptMgrByDept(String deptno) {
+		return sqlSession.selectOne(NS+"selectDeptMgrByDept",deptno);
+	}
+
+	@Override
+	public DepartmentVo selectDrafterDeptMgr(String id) {
+		return sqlSession.selectOne(NS+"selectDrafterDeptMgr", id);
 	}
 
 }
