+++ GreenLight/src/main/java/com/green/light/model/service/ApprovalServiceImpl.java
@@ -6,7 +6,10 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import com.green.light.model.mapper.IApprovalDao;
+import com.green.light.model.mapper.IDepartmentDao;
+import com.green.light.model.mapper.IDocumentDao;
 import com.green.light.vo.ApprJstreeVo;
+import com.green.light.vo.DepartmentVo;
 import com.green.light.vo.EmployeeVo;
 
 import lombok.extern.slf4j.Slf4j;
@@ -16,18 +19,11 @@ import lombok.extern.slf4j.Slf4j;
 public class ApprovalServiceImpl implements IApprovalService {
 	@Autowired
 	private IApprovalDao dao;
+	private IDepartmentDao deptdao;
 	
-	@Override
-	public EmployeeVo selectDrafterDeptMgr(String id) {
-		log.info("ApprovalServiceImpl selectDrafterDeptMgr 자동결재선 기안자의 부서장 조회 : {}", id);
-		return dao.selectDrafterDeptMgr(id);
-	}
 	
-	@Override
-	public EmployeeVo selectDeptMgrByDept(String deptno) {
-		log.info("ApprovalServiceImpl selectDeptMgrByDept 자동결재선 부서별 부서장 조회 : {}", deptno);
-		return dao.selectDeptMgrByDept(deptno);
-	}
+	
+	
 	
 	@Override
 	public EmployeeVo selectCEO() {
