+++ GreenLight/src/main/resources/sqls/ApprovalStatement.xml
@@ -13,6 +13,7 @@
 
 <resultMap type="com.green.light.vo.DepartmentVo" id="DepartmentVoMap">
 	<result property="deptno" column="DEPTNO"/>
+	<result property="dept_mgr" column="DEPT_MGR"/>
 </resultMap>
 
 <resultMap type="com.green.light.vo.CommonVo" id="CommonVoMap">
@@ -30,24 +31,9 @@
 </resultMap>
 
 
-<!-- 자동결재선 기안자의 부서장 조회 selectDrafterDeptMgr -->
-<select id="selectDrafterDeptMgr" resultMap="EmployeeVoMap">
-SELECT E2.ID, E2.NAME, C.CODE_NAME
-	FROM EMPLOYEE E
-	JOIN DEPARTMENT D ON E.DEPTNO = D.DEPTNO 
-	JOIN EMPLOYEE E2 ON D.DEPT_MGR = E2.ID 
-	JOIN COMMON C ON C.CODE = E2.SPOT AND C."TYPE" = '직위'
-	WHERE E.ID = #{id}
-</select>
 
-<!-- 자동결재선 부서별 부서장 조회 selectDeptMgrByDept -->
-<select id="selectDeptMgrByDept" resultMap="EmployeeVoMap">
-SELECT E.ID, E.NAME, C.CODE_NAME
-	FROM EMPLOYEE E
-	JOIN DEPARTMENT D ON E.DEPTNO = D.DEPTNO AND E."POSITION" = '01'
-	JOIN COMMON C ON C.CODE = E.SPOT AND C."TYPE" = '직위'
-	WHERE E.DEPTNO = #{deptno}
-</select>
+
+
 
 <!-- 자동결재선 대표이사 조회 selectCEO -->
 <select id="selectCEO" resultMap="EmployeeVoMap">
